feat:add basic UI
